Refactoring: Code Clean up in sprint class (addresses #275)

+++ src/main/java/switch2021/project/dataModel/JPA/ProjectJpa.java
@@ -23,7 +23,7 @@ public class ProjectJpa {
     private String businessSector;
     private String startDate;
     private int numberOfSprints;
-    private int sprintDuration;
+    private long sprintDuration;
     private double budget;
     private String endDate;
     private String typology;
